# backend/openapi.yaml
openapi: 3.1.0
info:
  title: ServeFlow API
  version: 1.0.0
  description: Existing ServeFlow REST API. Organization ownership is derived from the authenticated JWT, never from request query parameters.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
...
  /bookings:
    get:
      operationId: listBookings
      summary: List bookings
      description: "Defaults to page 1, 20 items, sorted by booking_date ascending. Sort fields: booking_date, created_at, status."
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/search"
        - name: status
          in: query
          schema:
            type: string
            enum: [BOOKED, CANCELLED, COMPLETED]
        - name: sort
          in: query
          schema:
            type: string
            enum: [booking_date, created_at, status]
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        "200":
          $ref: "#/components/responses/BookingList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createBooking
      summary: Create a booking
      requestBody:
        $ref: "#/components/requestBodies/Booking"
      responses:
        "201":
          description: Created booking.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /bookings/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      operationId: getBooking
      summary: Get a booking
      responses:
        "200":
          description: Booking.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateBooking
      summary: Update a booking
      requestBody:
        $ref: "#/components/requestBodies/BookingUpdate"
      responses:
        "200":
          description: Updated booking.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: cancelBooking
      summary: Cancel a booking
      responses:
        "204":
          description: Cancelled.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /appointments:
    get:
      operationId: listAppointments
      summary: List appointments
      description: "Defaults to page 1, 20 items, sorted by booking_date ascending. Sort fields: booking_date, created_at, status."
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/search"
        - name: status
          in: query
          schema:
            type: string
            enum: [BOOKED, CANCELLED, COMPLETED]
        - name: sort
          in: query
          schema:
            type: string
            enum: [booking_date, created_at, status]
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        "200":
          $ref: "#/components/responses/AppointmentList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createAppointment
      summary: Create an appointment and queue its notification/email event
      requestBody:
        $ref: "#/components/requestBodies/Appointment"
      responses:
        "201":
          description: Created appointment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /appointments/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      operationId: getAppointment
      summary: Get an appointment
      responses:
        "200":
          description: Appointment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateAppointment
      summary: Update an appointment
      requestBody:
        $ref: "#/components/requestBodies/AppointmentUpdate"
      responses:
        "200":
          description: Updated appointment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: cancelAppointment
      summary: Cancel an appointment
      responses:
        "204":
          description: Cancelled.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    page:
      name: page
      in: query
      description: Page number, from 1 through 1,000,000.
      schema:
        type: integer
        minimum: 1
        maximum: 1000000
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    search:
      name: search
      in: query
      schema:
        type: string
        maxLength: 100
...
  requestBodies:
...
    Booking:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BookingInput"
    BookingUpdate:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BookingUpdate"
    Appointment:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppointmentInput"
    AppointmentUpdate:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppointmentUpdate"
...
  schemas:
...
    BookingInput:
      type: object
      required: [customer_id, service_id, booking_date, start_time, end_time]
      unevaluatedProperties: false
      properties:
        customer_id:
          type: string
          format: uuid
        service_id:
          type: string
          format: uuid
        technician_id:
          type: string
          format: uuid
        booking_date:
          type: string
          format: date
          description: Local business date; no timezone is accepted.
        start_time:
          type: string
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
          description: Local 24-hour time in HH:MM format.
        end_time:
          type: string
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
          description: Local 24-hour time in HH:MM format, later than start_time.
        notes:
          type: string
          maxLength: 2000
    BookingUpdate:
      allOf:
        - $ref: "#/components/schemas/BookingInput"
        - type: object
          properties:
            status:
              type: string
              enum: [BOOKED, CANCELLED, COMPLETED]
    AppointmentInput:
      type: object
      required: [customer_id, service_id, appointment_date, start_time, end_time]
      unevaluatedProperties: false
      properties:
        customer_id:
          type: string
          format: uuid
        service_id:
          type: string
          format: uuid
        technician_id:
          type: string
          format: uuid
        appointment_date:
          type: string
          format: date
          description: Local business date; no timezone is accepted.
        start_time:
          type: string
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
          description: Local 24-hour time in HH:MM format.
        end_time:
          type: string
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
          description: Local 24-hour time in HH:MM format, later than start_time.
        notes:
          type: string
          maxLength: 2000
    AppointmentUpdate:
      allOf:
        - $ref: "#/components/schemas/AppointmentInput"
        - type: object
          properties:
            status:
              type: string
              enum: [BOOKED, CANCELLED, COMPLETED]
...
    Booking:
      allOf:
        - $ref: "#/components/schemas/BookingInput"
        - type: object
          required: [id, status]
          properties:
            id:
              type: string
              format: uuid
            status:
              type: string
              enum: [BOOKED, CANCELLED, COMPLETED]
    BookingResponse:
      type: object
      required: [booking]
      properties:
        booking:
          $ref: "#/components/schemas/Booking"
    Appointment:
      allOf:
        - $ref: "#/components/schemas/AppointmentInput"
        - type: object
          required: [id, status]
          properties:
            id:
              type: string
              format: uuid
            status:
              type: string
              enum: [BOOKED, CANCELLED, COMPLETED]
    AppointmentResponse:
      type: object
      required: [appointment]
      properties:
        appointment:
          $ref: "#/components/schemas/Appointment"
...
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
    Pagination:
      type: object
      required: [page, limit, total, total_pages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    BookingList:
      type: object
      required: [bookings, pagination]
      properties:
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        pagination:
          $ref: "#/components/schemas/Pagination"
    AppointmentList:
      type: object
      required: [appointments, pagination]
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
  responses:
    BadRequest:
      description: Invalid input.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required or token invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found in the authenticated organization.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Business-rule or uniqueness conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    BookingList:
      description: Booking list.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BookingList"
    AppointmentList:
      description: Appointment list.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppointmentList"
...